+++ insteon_mqtt/cmd_line/main.py
@@ -249,6 +249,10 @@ def parse_args(args):
     sp.add_argument("-r", "--reason", metavar="reason", type=str, default="",
                     help="Reason message to send with the command.  No "
                     "message with use 'scene'.")
+    sp.add_argument("-l", "--level", metavar="level", type=str, default=None,
+                    help="The brightness level to set the device if supported."
+                    " Otherwise on will be the defined scene level. Modem "
+                    "does not support the level command.")
     sp.add_argument("address", help="Device address or name.")
     sp.add_argument("group", help="Group (button) number of the "
                     "scene to trigger (use 1 for single buttons.) "