+++ tests/handler/test_ModemDbGet.py
@@ -64,6 +64,32 @@ class Test_ModemDbGet:
         assert db.device.sent[0]['handler'] == handler
         assert db.entry == test_entry
 
+    #-----------------------------------------------------------------------
+    def test_recs_not_used(self, caplog):
+        calls = []
+
+        def callback(success, msg, done):
+            calls.append(msg)
+
+        db = Mockdb()
+        db.device = MockDevice()
+        handler = IM.handler.ModemDbGet(db, callback)
+        proto = MockProtocol()
+
+        flags = Msg.DbFlags(False, True, False).to_bytes()[0]
+
+        b = bytes([0x02, 0x57,
+                   flags,  # flags
+                   0x01,  # group
+                   0x3a, 0x29, 0x84,  # addess
+                   0x01, 0x0e, 0x43])  # data
+        msg = Msg.InpAllLinkRec.from_bytes(b)
+        print(msg)
+        r = handler.msg_received(proto, msg)
+        assert r == Msg.FINISHED
+        assert isinstance(db.device.sent[0]['msg'], Msg.OutAllLinkGetNext)
+        assert db.device.sent[0]['handler'] == handler
+
 #===========================================================================