+++ tests/mqtt/test_Dimmer.py
@@ -382,7 +382,7 @@ class Test_Dimmer:
 
     #-----------------------------------------------------------------------
     def test_input_scene_reason(self, setup):
-        mdev, link, proto = setup.getAll(['mdev', 'link', 'proto'])
+        mdev, link, proto, dev = setup.getAll(['mdev', 'link', 'proto', 'dev'])
 
         qos = 2
         config = {'dimmer' : {
@@ -400,9 +400,16 @@ class Test_Dimmer:
 
         assert proto.sent[0].msg.cmd1 == 0x30
         assert proto.sent[0].msg.data[3] == 0x13
-        cb = proto.sent[0].handler.callback
-        assert cb.keywords == {"reason" : "ABC"}
+        cb = proto.sent[0].handler.on_done
+        assert dev.broadcast_reason == ""
+        # Signal a failure
+        cb(False, "Done", None)
+        assert dev.broadcast_reason == ""
+        # Signal a success
+        cb(True, "Done", None)
+        assert dev.broadcast_reason == "ABC"
         proto.clear()
+        dev.broadcast_reason = ""
 
         payload = b'{ "on" : "ON", "reason" : "DEF" }'
         link.publish(topic, payload, qos, retain=False)
@@ -410,9 +417,62 @@ class Test_Dimmer:
 
         assert proto.sent[0].msg.cmd1 == 0x30
         assert proto.sent[0].msg.data[3] == 0x11
-        cb = proto.sent[0].handler.callback
-        assert cb.keywords == {"reason" : "DEF"}
+        cb = proto.sent[0].handler.on_done
+        # Signal a success
+        cb(True, "Done", None)
+        assert dev.broadcast_reason == "DEF"
         proto.clear()
 
+
+    #-----------------------------------------------------------------------
+    def test_input_scene_level(self, setup):
+        mdev, link, proto, dev = setup.getAll(['mdev', 'link', 'proto', 'dev'])
+
+        qos = 2
+        config = {'dimmer' : {
+            'scene_topic' : 'foo/{{address}}/scene',
+            'scene_payload' : ('{ "cmd" : "{{json.on.lower()}}"'
+                               '{% if json.level is defined %}'
+                               ',"level" : "{{json.level}}"'
+                               '{% endif %} }')}}
+        mdev.load_config(config, qos=qos)
+
+        mdev.subscribe(link, qos)
+        topic = link.client.sub[2].topic
+
+        # just ON command
+        payload = b'{ "on" : "on"}'
+        link.publish(topic, payload, qos, retain=False)
+        assert len(proto.sent) == 1
+        assert proto.sent[0].msg.cmd1 == 0x30
+        assert proto.sent[0].msg.data[3] == 0x11  #cmd1
+        assert proto.sent[0].msg.data[1] == 0x00  #use_on_level
+        assert proto.sent[0].msg.data[2] == 0x00  #on_level
+        cb = proto.sent[0].handler.on_done
+        # Check default reason value if not specified
+        cb(True, "Done", None)
+        assert dev.broadcast_reason == IM.on_off.REASON_DEVICE
+        proto.clear()
+        proto.clear()
+
+        # just OFF command
+        payload = b'{ "on" : "off"}'
+        link.publish(topic, payload, qos, retain=False)
+        assert len(proto.sent) == 1
+        assert proto.sent[0].msg.cmd1 == 0x30
+        assert proto.sent[0].msg.data[3] == 0x13  #cmd1
+        assert proto.sent[0].msg.data[1] == 0x01  #use_on_level
+        assert proto.sent[0].msg.data[2] == 0x00  #on_level
+        proto.clear()
+
+        # just ON with level
+        payload = b'{ "on" : "on", "level": 128}'
+        link.publish(topic, payload, qos, retain=False)
+        assert len(proto.sent) == 1
+        assert proto.sent[0].msg.cmd1 == 0x30
+        assert proto.sent[0].msg.data[3] == 0x11  #cmd1
+        assert proto.sent[0].msg.data[1] == 0x01  #use_on_level
+        assert proto.sent[0].msg.data[2] == 0x80  #on_level
+        proto.clear()
 
 #===========================================================================